- Make the autosegs read-only. - Derive PClass from dobject.cpp. This has one major ramification: Since the PClass   is not allocated until runtime, you cannot initialize any static/global data   structures with pointers to PClasses using RUNTIME_CLASS. Attempting to do so   will just initialize with a NULL pointer. Instead, you can initialize using   the address of the pointer returned by RUNTIME_CLASS and dereference that. By   the time you have an opportunity to dereference it, it will no longer be NULL. - Sync CmakeLists.txt. - Random fixes for problems GCC spotted.
SVN r1852 (scripting)

+++ src/p_spec.cpp
@@ -1332,7 +1332,7 @@ static void P_SpawnScrollers(void)
 		if (special != 0)
 		{
 			int max = LineSpecialsInfo[special] != NULL ? LineSpecialsInfo[special]->map_args : countof(l->args);
-			for (int arg = max; arg < countof(l->args); ++arg)
+			for (int arg = max; arg < (int)countof(l->args); ++arg)
 			{
 				if (l->args[arg] != 0)
 				{
